Add voting result page and components
A page was created where the user can see which referee was assigned to
which match,after the voting process closed. This page has been added
to the app router and the user can access it using the navigation bar.
They can also see the final scores of each referee.